requirements.txt
appdirs==1.4.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.11
Django==1.10.5
django-cors-headers==2.0.2
django-filter==1.0.1
djangorestframework==3.5.3
enum34==1.1.6
ipython==5.2.2
ipython-genutils==0.1.0
Markdown==2.6.8
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
scandir==1.4
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
